// types/graphics.d.ts
export interface GraphicsState {
    renderLayers: RenderLayer[];
    camera: Camera;
    lighting: any;
}
export interface RenderLayer {
    id: string;
    sprites: Sprite[];
    visible: boolean;
}
export interface Sprite {
    id: string;
    texture: pc.Texture;
    position: pc.Vec3;
}
export interface Camera {
    position: pc.Vec3;
    target: pc.Vec3;
}